+++ doc/conf.py
@@ -251,6 +251,7 @@ man_pages = [("index", "cf_xarray", "cf_xarray Documentation", [author], 1)]
 intersphinx_mapping = {
     "python": ("https://docs.python.org/3/", None),
     "xarray": ("https://xarray.pydata.org/en/stable/", None),
+    "pandas": ("https://pandas.pydata.org/pandas-docs/stable", None),
 }
 
 autosummary_generate = True