+++ doc/whats-new.rst
@@ -5,7 +5,7 @@ What's New
 
 v 0.7.1 (unreleased)
 ====================
-- added :py:func:`encode_compress` and :py:func:`decode_compress` for writing pandas MultiIndex-es to file.
+- added encoder and decoder for writing pandas MultiIndex-es to file using "compression by gathering".
   See :ref:`coding` for more. By `Deepak Cherian`_.
 - added another type of vertical coordinate to decode: ``ocean_sigma_coordinate``. By `Kristen Thyng`_.